// src/generated/sdk/models/activityLogOut.ts
/**
 * Generated by orval v8.5.1 🍺
 * Do not edit manually.
 * QMS Backend
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityAction } from "./activityAction";
import type { ActivityLogOutMeta } from "./activityLogOutMeta";
import type { ActivityOutcome } from "./activityOutcome";
import type { EntityType } from "./entityType";

export interface ActivityLogOut {
  id: number;
  org_id: number | null;
  entity_type: EntityType;
  entity_id: number;
  action: ActivityAction;
  title: string;
  actor_id: number | null;
  actor_first_name: string | null;
  meta: ActivityLogOutMeta;
  created_at: string;
  outcome: ActivityOutcome;
  error_type: string | null;
  error_message: string | null;
  request_id: string | null;
}